pyproject.toml
[tool.poetry]
name = "convolucao-laplaciano"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "convolucao_laplaciano"}]

[tool.poetry.dependencies]
python = "^3.11"
polars = "^0.17.13"
plotly = "^5.14.1"
numpy = "^1.24.3"
pillow = "^9.5.0"
taskipy = "^1.10.4"
streamlit = "^1.22.0"
streamlit-toggle-switch = "^1.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
dashboard = "poetry run streamlit run convolucao_laplaciano/dashboard/Convolucao.py"
